archive/backend-2025-10-20/requirements.txt
requests==2.32.3
python-dotenv==1.0.1
google-generativeai==0.7.2
Pillow==10.4.0
fastapi==0.112.2
uvicorn[standard]==0.30.6
asyncpg==0.29.0
supabase==2.6.0
passlib==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
# Pin bcrypt to a version compatible with passlib to avoid __about__ attribute error
bcrypt==4.0.1
python-slugify
reportlab==3.6.13
matplotlib==3.8.4
openpyxl==3.1.5

